refactor extends relationship in GroupByDataSet
